API_doc/jwt/jwt_c.py
from datetime import datetime, timedelta, timezone
from jose import JWTError, jwt
import secrets


SECRET_KEY = secrets.token_urlsafe(32)
ALGORITHM = "HS256"


def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

    return encoded_jwt


def decode_access_token(encoded_jwt):
    return jwt.decode(encoded_jwt, SECRET_KEY, algorithms=[ALGORITHM])
