requirements.txt
catboost==1.2.2
fastapi==0.109.0
httpx==0.26.0
numpy==1.26.3
onnxruntime==1.16.3
passlib==1.7.4
pydantic==1.10.13
python_jose==3.3.0
SQLAlchemy==2.0.25
SQLAlchemy==2.0.23
SQLAlchemy_Utils==0.41.1
starlette==0.35.1
